refactor: 라우트 상수 처리 및 그외 스타일 수정
