+++ src/constants/routes.ts
@@ -1,0 +1,6 @@
+const PAGE_ROUTES = Object.freeze({
+  MAIN: '/',
+  CONFIRM: '/confirm',
+});
+
+export default PAGE_ROUTES;
